requirements.txt
Django==1.7.1
Markdown==2.5.1
MySQL-python==1.2.5
argparse==1.2.1
arrow==0.4.4
django-filter==0.8
djangorestframework==2.4.4
python-dateutil==2.2
simplifycommerce-sdk-python==1.1.1
six==1.8.0
wsgiref==0.1.2
